simplify server, db. separate client/server
Also get's rid of cukes!
