--- features/README.md
@@ -1,3 +1,0 @@
-This folder contains the feature specifications for this sidescroller
-game. The specifications are written according the gherkin standard.
-https://cucumber.io/docs/reference
