+++ src/app.js
@@ -1,4 +1,4 @@
-var server = require('../server/server').Server();
-var db = require('../server/db').Database();
+var server = require('./server/server');
+var db = require('./server/db');
 
 server.start(3000);
