+++ src/client/index.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-    <script src="hi.js"></script>
+    <script src="meatpocalypse.js"></script>
     <title>Hello world!</title>
     </head>
     <body>
